requirements.txt
asgiref==3.2.7
beautifulsoup4==4.8.2
certifi==2020.4.5.2
chardet==3.0.4
Django==3.0.5
django-modelcluster==5.0.2
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.11.0
draftjs-exporter==2.1.7
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
Jinja2==2.11.2
l18n==2018.5
MarkupSafe==1.1.1
Pillow==7.1.2
pytz==2019.3
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.1.1
urllib3==1.25.9
wagtail==2.9
webencodings==0.5.1
Willow==1.3
XlsxWriter==1.2.9
